requirements.txt
gunicorn
flask
flask_sqlalchemy
psycopg2-binary
werkzeug
python-dotenv
